+++ src/util/index.ts
@@ -128,3 +128,7 @@ export const extractEntityId = (id: string): Either<AlexaDeviceError, string> =>
     ),
     E.map(({ matches }) => matches![0]),
   );
+
+export const round = (value: number, decimals: number) => {
+  return Number(Math.round(Number(value + 'e' + decimals)) + 'e-' + decimals);
+};